spinner over canvasses, thresholdchecking, custom colors for canvas, empty json config will set from defaults
There will be a spinner over the canvas when data is (re)loading, empty charts and charts with thresholds will be checked. Depending on data values, the chartcanvas will be painted in a different color. When some configsettings aren't given, they will be filled automatically with given defaults (from another local .JSON file)